// demo-mcp-21-code-generator-node/input/reference-repos/MSProdOffeLoad/src/common/utils/enums/taks.enum.ts
/**
 * Enumeraciones usadas internamente para la trazabilidad y descripción de tareas en los logs del microservicio
 * @author Oscar Avila
 */

export enum Etask {
    ERROR_PROVIDER = 'ERROR PROVIDER',
    UPLOAD_FILE = 'SUBIENDO ARCHIVO',
    FILE_UPLOADED = 'ARCHIVO CARGADO',
    READ_FILE = 'LEYENDO ARCHIVO',
    END_CONNECTION = 'FIN DE LA CONEXION',
    CREATE = 'PROCESS_CREATE',
    UPDATE = 'ACTUALIZANDO_MOCKUP',
    APM='CAPTURE_APM_QUERY',
    LOAD_MESSAGE = 'CARGANDO_MENSAJES',
    ERROR_LOAD_PARAM = 'ERROR_CARGANDO_PARAMETROS',
    CHANNEL = 'VALIDATE_CHANNEL',
    CONSUMED_SERVICE = 'CONSUMED_SERVICE',
    PROCESS_KAFKA_ERROR_EVENT = 'EVENTO KAFKA INCORRECTO',
    KAFKA_ERROR_EVENT = 'ERROR EVENTO KAFKA',
    PROCESS_KAFKA_EVENT = 'EVENTO ACEPTADO',
    FINISH_KAFKA_EVENT = 'FINISH ACEPTADO',
    VALIDATE_REQUEST='VALIDATE_REQUEST',
    DATALOAD_MANUAL = 'GENERANDO DATALOAD MANUAL',
    DATALOAD = 'GENERANDO DATALOAD',
    FIND_EQUIPMENT = 'CONSULTANDO CATALOGOS TERMINALES',
    FIND_TECHNOLOGY = 'CONSULTANDO CATALOGOS TECNOLOGÍA',
    FIND_POSPAGO = 'CONSULTANDO CATALOGOS POSPAGO',
    FIND_PREPAGO = 'CONSULTANDO CATALOGOS PREPAGO',
    FIND_HOMES = 'CONSULTANDO CATALOGOS HOGARES',
    FIND_DISPONIBILITY = 'CONSULTANDO DISPONIBILIDAD',
    FIND_PRICES = 'CONSULTANDO PRECIOS',
    FIND_LIST_PARENTS = 'CONSULTANDO PADRES-HIJOS',
    FIND_DATA = 'CONSULTANDO INFORMACIÓN',
    SAVE_LIST_PARENTS = 'GUARDANDO PADRES-HIJOS',
    GENERATE_CSV = 'GENERANDO ARCHIVO CSV',
    GET_IMAGES = 'OBTENIENDO IMAGENES',
    CREATE_PRODUCT_DATA = 'CREANDO PRODUCT-DATA',
    CREATE_NAME_FOR_TECHNOLOGY = 'CREATE NAME FOR TECHNOLOGY',
    CREATE_SALES_CATALOG = 'CREANDO SALES-CATALOG',
    CREATE_ATTRIBUTES_PRODUCT = 'CREANDO ATTRIBUTES-PRODUCTS',
    CREATE_ATTRIBUTES_DICTIONARY = 'CREANDO ATTRIBUTES-DICTIONARY',
    CREATE_ATTACHMENT_DATA = 'CREANDO ATTACHMENT-DATA',
    CREATE_PRODUCT_INVENTORY = 'CREANDO PRODUCT-INVENTORY',
    CREATE_PRICE_LIST = 'CREANDO PRICE-LIST',
    START_MAP_PRICES = 'INICIA EL REGISTRO DE PRECIOS',
    END_MAP_PRICES = 'FIN DEL REGISTRO DE PRECIOS',
    REQUEST_HTTP = 'REQUEST_HTTS',
    ERROR_TRACE = 'ERROR IN TRACEABILITY',
    START_CREATE_DATALOAD = 'INICIA CREACIÓN DATALOAD',
    GET_DATA = 'CONSULTANDO INFORMACIÓN DE PRODUCTOS',
    GET_PARENT_FROM_FEATURE = 'CONSULTANDO AGRUPACIÓN',
    FINISH_CREATE_DATALOAD = 'FINALIZA CREACIÓN DATALOAD'
}

//Descripción de las tareas que se realizan en el microservicio
export enum ETaskDesc {
    CHANNEL = 'Validation of the channel',
    ERROR_LOAD_PARAM = 'Error cargando parametros',
    CONSUMED_SERVICE = 'Result service',
    ERROR_GENERATE_DATALOAD = 'Error generando dataload',
    ERROR_KAFKA_EVENT = 'Error orquestando dataload kafka',
    INVALID_DATALOAD = 'Dataload invalido',
    INVALID_CATEGORY = 'Categoria invalida',
    ERROR_FIND_DATA = 'Error consultando información',
    ERROR_GROUP_PRODUCTS = 'Error agrupando productos',
    ERROR_GENERATE_PARENTS = 'Error generando padres',
    ERROR_CREATE_NAME_FOR_TECHNOLOGY = 'Error creando el nombre para tecnologia',
    ERROR_GENERATE_ROWS = 'Error generando filas',
    ERROR_CREATE_FILE = 'Error creando archivo',
    ERROR_SAVE_LIST_PARENTS = 'Error guardando lista de padres',
    ERROR_GET_LIST_PARENTS = 'Error consultando lista de padres',
    INVALID_IMAGES = 'No se encontraron imagenes',
    INVALID_DATA = 'No se encontro información',
    ERROR_SFTP = 'Error guardando archivo sftp',
    ERROR_FIND_IMAGES = 'Error consultando imagenes',
    KAFKA_EVENT_FEATURES = 'Procesando topic CalendarAlarmFeaturesTopic',
    KAFKA_EVENT_PRICES = 'Procesando topic CalendarAlarmPricesTopic',
    KAFKA_EVENT_DISPONIBILITY = 'Procesando topic CalendarAlarmDisponibilityTopic',
    START_EVENT = 'Inicia ejecucion por evento',
    FINISH_EVENT = 'Finaliza ejecucion por evento'
}

export enum ETaskTrace {
    FAILED ="FAILED",
    SUCCESS="SUCCESS"
}